requirements.txt
aiohttp==3.10.5
asyncpg==0.30.0
python-dotenv==1.0.1
APScheduler==3.10.4
beautifulsoup4==4.12.3
playwright>=1.55.0,<2.0.0